Revert "Extract pusher to lib"
