+++ api/services/maintenance-service/src/notifier/pusher_chanels.js
@@ -6,12 +6,5 @@ module.exports.PusherChannels = {
         events: {
             auto_update_schedule_updated: 'auto_update_schedule_updated'
         }
-    },
-    screens: {
-        name: 'screens',
-        events: {
-            update: 'update',
-            refresh: 'refresh'
-        }
     }
-};
+}
