+++ api/services/maintenance-service/src/screen_notifications/refresh_screens.js
@@ -1,13 +1,8 @@
 'use strict';
 
+const Notifier = require('../notifier/notifier');
 
 const ModulePathManager = require('../module_path_manager');
-const Notifier = require(ModulePathManager.getBasePath() + 'lib/notifier/notifier');
-const PusherChannels = require(ModulePathManager.getBasePath() + 'lib/notifier/pusher_channels').PusherChannels;
-
-const ScreensChannel = PusherChannels.screens;
-const ChannelEvent = ScreensChannel.events.refresh;
-
 const responseHelper = require(ModulePathManager.getBasePath() + 'lib/helpers/http_response_helper');
 
 
@@ -25,5 +20,5 @@ module.exports.refresh = (event, context, callback) => {
 };
 
 function sendRefreshSignal(screens) {
-    Notifier.pushNotification(ScreensChannel.name, ChannelEvent, {screens: screens});
+    Notifier.pushNotification('screens', 'refresh', {screens: screens});
 }
